refactor(checker): extract FlowDeferMemos into control_flow/core/defer_classification
Move the per-walk CALL narrow/divert classification memo (FlowDeferMemos
struct + call_node_may_narrow_or_divert_cached wrapper) out of
flow/control_flow/core.rs into a dedicated core/defer_classification.rs
submodule. Pure code motion, no behavior change: keeps core.rs under its
recorded size ratchet (1825 <= 1848) without raising the baseline.

The witness test's structural guard now reads the classifier and memo type
from the new submodule path.
