+++ crates/tsz-checker/src/flow/control_flow/core.rs
@@ -80,8 +80,11 @@ pub(crate) const fn is_session_stable_flow_cache_symbol(symbol: SymbolId) -> boo
     symbol.0 & (FLOW_CACHE_SYNTHETIC_BIT | FLOW_CACHE_PER_NODE_BIT) != FLOW_CACHE_SYNTHETIC_BIT
 }
 
+mod defer_classification;
 mod flow_query;
 mod flow_traversal;
+
+pub(crate) use defer_classification::FlowDeferMemos;
 
 #[must_use]
 pub(crate) fn flow_cache_entries(cache: &FlowCache) -> usize {
@@ -211,20 +214,6 @@ const fn flow_step_budget(flow_node_count: usize) -> usize {
     } else {
         scaled
     }
-}
-
-/// Per-walk classification memos shared by the linear-passthrough chase and the
-/// defer classifier. Both decisions are pure functions of a flow node within a
-/// single backward walk (`reference`/`symbol_id` are fixed and the type / type-
-/// predicate caches are immutable mid-walk), so each is computed at most once per
-/// node per walk. Bundling them keeps the chase within its argument budget while
-/// the recursion in `antecedent_requires_defer` reuses both tables.
-#[derive(Default)]
-pub(super) struct FlowDeferMemos {
-    /// `antecedent_requires_defer` result keyed by flow-node id.
-    pub(super) defer: FxHashMap<FlowNodeId, bool>,
-    /// `call_node_may_narrow_or_divert` result keyed by flow-node id.
-    pub(super) call_divert: FxHashMap<FlowNodeId, bool>,
 }
 
 /// Find a symbol's representative identifier node, preferring a usage site over
@@ -1473,31 +1462,6 @@ impl<'a> FlowAnalyzer<'a> {
         };
         self.predicate_signature_for_type(callee_type)
             .is_some_and(|sig| sig.predicate.asserts)
-    }
-
-    /// Per-walk memoized wrapper around [`Self::call_node_may_narrow_or_divert`].
-    ///
-    /// The classification is a pure function of the CALL flow node within a single
-    /// backward walk (it reads the immutable per-check type cache and resolved
-    /// type-predicate tables; `reference`/`symbol_id` do not enter it). Both the
-    /// linear-passthrough chase and the defer classifier re-derive it for the same
-    /// node many times per walk — the chase alone re-scans overlapping pass-through
-    /// runs on every worklist pop — so without a memo a call-dense scope pays the
-    /// full predicate-signature extraction (`classify_for_predicate_signature`,
-    /// `callable_shape`) thousands of times per reference read. Caching by flow-node
-    /// id collapses that to one extraction per node per walk with no change in value.
-    pub(super) fn call_node_may_narrow_or_divert_cached(
-        &self,
-        flow_id: FlowNodeId,
-        ant_flow: &FlowNode,
-        memo: &mut FxHashMap<FlowNodeId, bool>,
-    ) -> bool {
-        if let Some(&cached) = memo.get(&flow_id) {
-            return cached;
-        }
-        let result = self.call_node_may_narrow_or_divert(ant_flow);
-        memo.insert(flow_id, result);
-        result
     }
 
     /// Helper function for call handling in iterative mode.
